ADD: Custom video player

+++ controller/videoController.js
@@ -38,10 +38,11 @@ export const postUpload = async(req, res) => {
     const newVideo = await Video.create({
         fileUrl: path,
         title,
-        description
+        description,
+        creator: req.user.id
     })
-    console.log(newVideo);
-    // To Do : Upload and Save Video
+    req.user.videos.push(newVideo.id);
+    req.user.save();
     res.redirect(routes.videoDetail(newVideo.id))
 }
 
@@ -50,21 +51,27 @@ export const videoDetail = async(req, res) => {
         params: { id }
     } = req;
     try {
-        const video = await Video.findById(id)
+        const video = await Video.findById(id).populate("creator");
         res.render("videoDetail", { pageTitle: video.title, video});
+        console.log(video);
     } catch(error){
         console.log(error);
         res.redirect(routes.home);
     }
 }
-export const getEditVideo = async(req, res) => {
+export const getEditVideo = async (req, res) => {
     const {
         params: { id }
     } = req;
     try {
-        console.log(req);
         const video = await Video.findById(id);
-        res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+        console.log(typeof(video.creator));
+        console.log(typeof(req.user.id));
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+        } else{
+            res.render("editVideo", { pageTitle: `Edit ${video.title}`, video});
+        }
     } catch(error) {
         console.log(error);
         res.redirect(routes.home);
@@ -89,7 +96,12 @@ export const deleteVideo = async(req, res) => {
         params: { id }
     } = req;
     try {
-        await Video.findOneAndRemove({_id: id})
+        const video = await Video.findById(id);
+        if(String(video.creator) !== req.user.id){
+            throw Error();
+        } else{
+            await Video.findOneAndRemove({_id: id})
+        }
     } catch(error){
         console.log(error);
     }
